Move Button component to src/basics
